move logic to hoc
