+++ src/App.js
@@ -5,7 +5,7 @@ import Form from './components/Form'
 import Chat from './components/Chat'
 import Controls from './components/Controls'
 import VideoPlayer from './components/VideoPlayer'
-import youtubeFetcher from './utils/youtubeFetcher'
+import withSocket from './utils/socketSubscriber'
 
 const Global = createGlobalStyle`
   html, body, #root {
@@ -118,225 +118,11 @@ const getListStyle = isDraggingOver => ({
   padding: '1rem'
 });
 
-const reorder = (list, startIndex, endIndex) => {
-  const result = Array.from(list);
-  const [removed] = result.splice(startIndex, 1);
-  result.splice(endIndex, 0, removed);
-
-  return result;
-};
-
 class App extends React.Component {
-  constructor(props) {
-    super(props)
-    this.state = {
-      messages: [],
-      chat: [],
-      role: "pending"
-    }
-    this.makeSocket = this.makeSocket.bind(this)
-    this.appendToMessages = this.appendToMessages.bind(this)
-    this.appendToChat = this.appendToChat.bind(this)
-    this.handleVideoEnd = this.handleVideoEnd.bind(this)
-    this.handlePresent = this.handlePresent.bind(this)
-    this.onDragEnd = this.onDragEnd.bind(this)
-    this.goBack = this.goBack.bind(this)
-    this.pauseOrPlay = this.pauseOrPlay.bind(this)
-    this.goForward = this.goForward.bind(this)
-  }
-
-  async componentDidMount() {
-    console.log("Mounting...")
-    this.makeSocket()
-    
-    window.addEventListener('beforeunload', () =>{    
-      if (this.state.role === "presenter") this.state.socket.send(JSON.stringify({
-        intent: "presenter leaving"
-      }))
-    })
-
-    document.addEventListener("visibilitychange", () => {
-      if (document.visibilityState === 'visible') {
-        console.log("Checking socket health...")
-        if (this.state.socket.readyState !== WebSocket.OPEN && this.state.socket.readyState !== WebSocket.CONNECTING) {
-          console.log("Remounting, websocket was ", this.state.socket.readyState)
-          this.state.socket.close()
-          this.makeSocket()
-        }
-      }
-    })
-  }
-
-  makeSocket() {
-    let socket = new WebSocket("wss://connect.websocket.in/spec?room_id=queue-" + (window.location.pathname.slice(1) || 1))
-    socket.addEventListener('message', message => {
-      let parsed = JSON.parse(message.data)
-      if (parsed.intent === "look for presenter"
-        && this.state.role === "presenter") {
-        socket.send(JSON.stringify({
-          intent: "presenter exists",
-          messages: this.state.messages,
-        }))
-      }
-      else if (parsed.intent === "look for presenter"
-        && this.state.role === "no presenter") {
-        socket.send(JSON.stringify({
-          intent: "no presenter",
-          messages: this.state.messages,
-        }))
-      }
-      else if (parsed.intent === "presenter exists") {
-        this.setState({
-          messages: parsed.messages,
-          role: "client"
-        })
-      }
-      else if (parsed.intent === "no presenter") {
-        this.setState({
-          messages: parsed.messages,
-          role: "no presenter"
-        })
-      }
-      else if (parsed.intent === "presenter leaving") {
-        this.setState({
-          role: "no presenter"
-        })
-      }
-      else if (parsed.intent === "messages changed") {
-        this.setState({
-          messages: parsed.messages
-        })
-      }
-      else if (parsed.intent === "message added") {
-        this.appendToMessages(parsed.message)
-      }
-      else if (parsed.intent === "chat added") {
-        this.appendToChat(parsed.chat)
-      }
-    })
-    socket.addEventListener('open', () => {
-      socket.send(JSON.stringify({intent: "look for presenter"}))
-      this.setState({
-        socket
-      })
-    })
-    setTimeout(() => {
-      if (this.state.role === "pending") {
-        socket.send(JSON.stringify({
-          intent: "presenter exists",
-          messages: this.state.messages,
-        }))
-        this.setState({ role: "presenter" })
-      }
-    }, 3000)
-    return socket
-  }
-
-  componentWillUnmount() {
-    this.state.socket.close()
-  }
-
-  async appendToMessages(url) {
-    // let id = youtubeFetcher.urlToID(message)
-    // let name = await youtubeFetcher.idToName(id)
-    let name = await youtubeFetcher.urlToName(url)
-    let time = new Date().getTime() // play the same video twice in a row
-    if (name) this.setState(state => {
-      return {
-        messages: state.messages.concat([{url, name, time}])
-      }
-    })
-  }
-
-  async appendToChat(chat) {
-    chat.travelTime = 10 * Math.random() + 10
-    this.setState(state => {
-      return {
-        chat: state.chat.concat([chat])
-      }
-    }, () => {
-      setTimeout(() => {
-        this.setState(state => {
-          let newChat = state.chat.filter(c => c.time !== chat.time)
-          return {
-            chat: newChat
-          }
-        })
-      }, chat.travelTime * 1000 * 1.5) // Grace
-    })
-  }
-
-  handleVideoEnd() {
-    this.setState(state => {
-      return {
-        messages: state.messages.slice(1)
-      }
-    }, () => {
-      this.state.socket.send(JSON.stringify({
-        intent: "messages changed",
-        messages: this.state.messages
-      }))
-    })
-  }
-
-  handlePresent() {
-    this.state.socket.send(JSON.stringify({
-      intent: "presenter exists",
-      messages: this.state.messages
-    }))
-    this.setState({
-      role: "presenter"
-    })
-  }
-
-  goBack(e) {
-    if (e.detail) e.currentTarget.blur()
-    this.state.socket.send(JSON.stringify({
-      intent: "go back"
-    }))
-  }
-
-  pauseOrPlay(e) {
-    if (e.detail) e.currentTarget.blur()
-    this.state.socket.send(JSON.stringify({
-      intent: "pause or play"
-    }))
-  }
-
-  goForward(e) {
-    if (e.detail) e.currentTarget.blur()
-    this.state.socket.send(JSON.stringify({
-      intent: "go forward"
-    }))
-  }
-
-  onDragEnd(result) {
-    // dropped outside the list
-    let messages
-    if (!result.destination) {
-      messages = Array.from(this.state.messages)
-      messages.splice(result.source.index, 1)
-    }
-    else messages = reorder(
-      this.state.messages,
-      result.source.index,
-      result.destination.index
-    );
-
-    this.state.socket.send(JSON.stringify({
-      intent: "messages changed",
-      messages
-    }))
-
-    this.setState({
-      messages
-    });
-  }
-
   render() {
-    return <DragDropContext onDragEnd={this.onDragEnd}>
+    return <DragDropContext onDragEnd={this.props.onDragEnd}>
       {
-        this.state.chat.map(chat =>
+        this.props.chat.map(chat =>
           <ChatMessage key={chat.time} travelTime={chat.travelTime} y={chat.y}>
             {chat.message}
           </ChatMessage>
@@ -346,30 +132,30 @@ class App extends React.Component {
       <Container>
         <div>
           <ActionsContainer>
-            <PresenterButton>{this.state.role === "pending" ? "CLIENT" : this.state.role.toUpperCase()}</PresenterButton>
+            <PresenterButton>{this.props.role === "pending" ? "CLIENT" : this.props.role.toUpperCase()}</PresenterButton>
             {
-              this.state.role !== "presenter" &&
-              this.state.role !== "no presenter" &&
-              this.state.messages[0] &&
+              this.props.role !== "presenter" &&
+              this.props.role !== "no presenter" &&
+              this.props.messages[0] &&
               <Controls
-                goBack={this.goBack}
-                pauseOrPlay={this.pauseOrPlay}
-                goForward={this.goForward}
+                goBack={this.props.goBack}
+                pauseOrPlay={this.props.pauseOrPlay}
+                goForward={this.props.goForward}
               />
             }
             {
-              this.state.role === "no presenter" && <div><StyledButton onClick={this.handlePresent}>Present</StyledButton></div>
+              this.props.role === "no presenter" && <div><StyledButton onClick={this.props.handlePresent}>Present</StyledButton></div>
             }
           </ActionsContainer>
           {
-            this.state.messages.length > 0 && <Droppable droppableId="droppable">
+            this.props.messages.length > 0 && <Droppable droppableId="droppable">
               {(provided, snapshot) => (
                   <div
                     {...provided.droppableProps}
                     ref={provided.innerRef}
                     style={getListStyle(snapshot.isDraggingOver)}
                   >
-                    {this.state.messages.map((message, index) => (
+                    {this.props.messages.map((message, index) => (
                       <Draggable key={message.time} draggableId={message.time} index={index}>
                         {(provided, snapshot) => (
                           <div
@@ -394,19 +180,19 @@ class App extends React.Component {
             </Droppable>
           }
           <FormsContainer>
-            <Form socket={this.state.socket} appendToMessages={this.appendToMessages}/>
+            <Form socket={this.props.socket} appendToMessages={this.props.appendToMessages}/>
             <Divider/>
-            <Chat socket={this.state.socket} appendToChat={this.appendToChat}/>
+            <Chat socket={this.props.socket} appendToChat={this.props.appendToChat}/>
           </FormsContainer>
         </div>
       {
-        this.state.role === "presenter" &&
-        this.state.messages[0] &&
+        this.props.role === "presenter" &&
+        this.props.messages[0] &&
         <VideoPlayer
-          videoUrl={this.state.messages[0].url}
-          time={this.state.messages[0].time}
-          handleVideoEnd={this.handleVideoEnd}
-          socket={this.state.socket}
+          videoUrl={this.props.messages[0].url}
+          time={this.props.messages[0].time}
+          handleVideoEnd={this.props.handleVideoEnd}
+          socket={this.props.socket}
         />
       }
       </Container>
@@ -414,4 +200,4 @@ class App extends React.Component {
   }
 }
 
-export default App;
+export default withSocket(App);
